eval/pyproject.toml
# RAG Evaluation Framework
# Install: uv pip install -e eval/

[project]
name = "rust-local-rag-eval"
version = "0.1.0"
description = "Evaluation framework for rust-local-rag RAG system"
requires-python = ">=3.11"

# MVP dependencies (Phase 1)
dependencies = [
    "pyyaml>=6.0",       # Config loading
    "requests>=2.31",    # HTTP client for RAG API
    "rich>=13.0",        # Pretty CLI output
    "typer>=0.9",        # CLI framework
]

[project.optional-dependencies]
# Phase 2: A/B testing with statistics
phase2 = [
    "numpy>=1.24",       # Bootstrap CI, array ops
    "scipy>=1.11",       # Statistical tests (t-test)
]

# Phase 3: Visualization and advanced metrics
phase3 = [
    "pandas>=2.0",       # Data manipulation
    "matplotlib>=3.7",   # Plotting
    "seaborn>=0.12",     # Statistical visualization
]

# All optional dependencies
all = [
    "numpy>=1.24",
    "scipy>=1.11",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
]

[project.scripts]
rag-eval = "eval.run:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
